Initial commit + more.
This includes the basics of an initial project (including a .gitignore, README.md, and basic project files). I have also implemented the initial basics of the application - a screen that fetches weather stations from the API and displays their names in a table view. I have implemented one test class in this initial commit.